+++ src/_DataStructures_/DoublyLinkedList/Node.js
@@ -1,0 +1,8 @@
+class Node {
+  constructor(data, previous, next) {
+    this.data = data;
+    this.previous = previous;
+    this.next = next;
+  }
+}
+module.exports = Node;
