+++ src/_DataStructures_/DoublyLinkedList/index.js
@@ -1,11 +1,5 @@
 /* eslint-disable class-methods-use-this */
-class Node {
-  constructor(data, previous, next) {
-    this.data = data;
-    this.previous = previous;
-    this.next = next;
-  }
-}
+const Node = require('./Node');
 
 class DoublyLinkedList {
   constructor() {
